Fit the copula model

fit_copula fits the copula model.

Usage

fit_copula(
  sce,
  assay_use,
  input_data,
  new_covariate = NULL,
  marginal_list,
  family_use,
  copula = "vine",
  DT = TRUE,
  pseudo_obs = FALSE,
  epsilon = 1e-06,
  family_set = c("gaussian", "indep"),
  n_cores
)

Arguments

sce

A SingleCellExperiment object.

assay_use

A string which indicates the assay you will use in the sce. Default is 'counts'.

input_data

The input data.

new_covariate

A data.frame which contains covaraites of targeted simulated data from construct_data.

marginal_list

A list of fitted regression models from fit_marginal.

family_use

A string or a vector of strings of the marginal distribution. Must be one of 'poisson', 'nb', 'zip', 'zinb' or 'gaussian'.

copula

A string of the copula choice. Must be one of 'gaussian' or 'vine'. Default is 'vine'.

DT

A logic variable. If TRUE, perform the distributional transformation to make the discrete data 'continuous'. This is useful for discrete distributions (e.g., Poisson, NB). Default is TRUE.

pseudo_obs

A logic variable. If TRUE, use the empirical quantiles instead of theoretical quantiles for fitting copula. Default is FALSE.

epsilon

A numeric variable for preventing the transformed quantiles to collapse to 0 or 1.

family_set

A string or a string vector of the bivarate copula families. Default is c("gaussian", "indep").

n_cores

An integer. The number of cores to use.

Value

A list with the components:

new_mvu

A matrix of the new multivariate uniform distribution from the copula.

copula_list

A list of the fitted copula model. If using Gaussian copula, a list of correlation matrices; if vine, a list of vine objects.

model_aic

A vector of the marginal AIC and the copula AIC.

model_bic

A vector of the marginal BIC and the copula BIC.

Details

This function takes the result from fit_marginal as the input and and fit the copula model on the residuals.