Functions from gamlss/gamlss.add with bugs fixed

The control for ga(). From gamlss.add::ga.control() and gamlss::gam().

Usage

ga.control(
  offset = NULL,
  method = "REML",
  optimizer = c("outer", "newton"),
  control = list(),
  scale = 0,
  select = FALSE,
  knots = NULL,
  sp = NULL,
  min.sp = NULL,
  H = NULL,
  gamma = 1,
  paraPen = NULL,
  in.out = NULL,
  drop.unused.levels = TRUE,
  drop.intercept = NULL,
  discrete = FALSE,
  ...
)

Arguments

offset

The offset in the formula.

method

The smoothing parameter estimation method.

optimizer

An array specifying the numerical optimization method to use to optimize the smoothing parameter estimation criterion (given by method)

control

A list of fit control parameters to replace defaults returned by gam.control.

scale

If this is positive then it is taken as the known scale parameter. Negative signals that the scale parameter is unknown. 0 signals that the scale parameter is 1 for Poisson and binomial and unknown otherwise.

select

If this is TRUE then gam() can add an extra penalty to each term so that it can be penalized to zero.

knots

This is an optional list containing user specified knot values to be used for basis construction.

sp

A vector of smoothing parameters can be provided here.

min.sp

Lower bounds can be supplied for the smoothing parameters.

H

A user supplied fixed quadratic penalty on the parameters of the GAM can be supplied, with this as its coefficient matrix.

gamma

Increase this beyond 1 to produce smoother models.

paraPen

Optional list specifying any penalties to be applied to parametric model terms.

in.out

Optional list for initializing outer iteration.

drop.unused.levels

By default unused levels are dropped from factors before fitting. For some smooths involving factor variables you might want to turn this off.

drop.intercept

Set to TRUE to force the model to really not have the a constant in the parametric model part, even with factor variables present. Can be vector when formula is a list.

discrete

Experimental option for setting up models for use with discrete methods employed in bam.

...

Other arguments

Value

A control object

ga.control

NA